[#78] V2 merge

+++ src/main/java/project/newchat/chatroom/service/ChatRoomServiceImpl.java
@@ -1,25 +1,46 @@
 package project.newchat.chatroom.service;
 
 
-import java.time.LocalDateTime;
+import static java.time.LocalDateTime.now;
+import static project.newchat.common.type.ErrorCode.ALREADY_JOIN_ROOM;
+import static project.newchat.common.type.ErrorCode.FAILED_GET_LOCK;
+import static project.newchat.common.type.ErrorCode.INVALID_REQUEST;
+import static project.newchat.common.type.ErrorCode.NEED_TO_PASSWORD;
+import static project.newchat.common.type.ErrorCode.NONE_ROOM;
+import static project.newchat.common.type.ErrorCode.NOT_FOUND_HEART;
+import static project.newchat.common.type.ErrorCode.NOT_FOUND_ROOM;
+import static project.newchat.common.type.ErrorCode.NOT_FOUND_USER;
+import static project.newchat.common.type.ErrorCode.NOT_ROOM_CREATOR;
+import static project.newchat.common.type.ErrorCode.NOT_ROOM_MEMBER;
+import static project.newchat.common.type.ErrorCode.REQUEST_SAME_AS_CURRENT_TITLE;
+import static project.newchat.common.type.ErrorCode.ROOM_PASSWORD_MISMATCH;
+
 import java.util.ArrayList;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
+import org.apache.kafka.clients.admin.NewTopic;
+import org.redisson.api.RLock;
+import org.redisson.api.RedissonClient;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import project.newchat.chatmsg.repository.ChatMsgRepository;
 import project.newchat.chatroom.controller.request.ChatRoomRequest;
+import project.newchat.chatroom.controller.request.ChatRoomUpdateRequest;
 import project.newchat.chatroom.domain.ChatRoom;
 import project.newchat.chatroom.dto.ChatRoomDto;
+import project.newchat.chatroom.dto.ChatRoomUserDto;
 import project.newchat.chatroom.repository.ChatRoomRepository;
 import project.newchat.common.exception.CustomException;
 import project.newchat.common.type.ErrorCode;
+import project.newchat.heart.domain.Heart;
+import project.newchat.heart.repository.HeartRepository;
 import project.newchat.user.domain.User;
 import project.newchat.user.repository.UserRepository;
 import project.newchat.userchatroom.domain.UserChatRoom;
@@ -37,73 +58,138 @@ public class ChatRoomServiceImpl implements ChatRoomService {
 
   private final UserChatRoomRepository userChatRoomRepository;
 
+  private final HeartRepository heartRepository;
+
+  private final RedissonClient redissonClient;
+
+  private final KafkaTemplate<String, String> kafkaTemplate;
+  private final String BASIC_TOPIC = "CHAT_ROOM";
+
   @Override
   @Transactional
-  public void createRoom(ChatRoomRequest chatRoomRequest, Long userId) {
+  public ChatRoom createRoom(ChatRoomRequest chatRoomRequest, Long userId) {
     // 유저정보조회
-    User findUser = getFindUser(userId);
+    User findUser = getUser(userId);
+    String password = chatRoomRequest.getPassword();
     // chatroom 생성
     ChatRoom chatRoom = ChatRoom.builder()
         .roomCreator(findUser.getId())
         .title(chatRoomRequest.getTitle())
         .userCountMax(chatRoomRequest.getUserCountMax())
-        .createdAt(LocalDateTime.now())
-        .updatedAt(LocalDateTime.now())
+        .createdAt(now())
+        .updatedAt(now())
+        .isPrivate(false)
         .build();
+    if (password != null) {
+      chatRoom.setPassword(password);
+      chatRoom.setIsPrivate(true);
+    }
     ChatRoom save = chatRoomRepository.save(chatRoom);
 
     // 연관관계 user_chat room 생성
     UserChatRoom userChatRoom = UserChatRoom.builder()
         .user(findUser)
         .chatRoom(save)
+        .joinDt(now())
         .build();
+    System.out.println("---------------------------------");
+    System.out.println("userChatRoom = " + userChatRoom.getUser());
+    System.out.println("userChatRoom = " + userChatRoom.getChatRoom());
+    System.out.println("---------------------------------");
     // save
     userChatRoomRepository.save(userChatRoom);
+    String topicName = BASIC_TOPIC + save.getId();
+    NewTopic newTopic = new NewTopic(topicName, 1, (short) 1);
+    // Kafka Topic에 구독자 추가
+    kafkaTemplate.send(newTopic.name(), "Subscribed");
+    return save;
   }
 
   @Override
   @Transactional
-  public void joinRoom(Long roomId, Long userId) {
-    // 유저 조회
-    User findUser = getFindUser(userId);
+  public void joinRoom(Long roomId, Long userId, ChatRoomRequest chatRoomRequest) {
+    RLock lock = redissonClient.getLock("joinRoomLock:" + roomId);
+    try {
+      boolean available = lock.tryLock(1, TimeUnit.SECONDS);
 
-    // room 조회
-    ChatRoom chatRoom = chatRoomRepository.findById(roomId)
-        .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_ROOM));
+      if (!available) {
+        throw new CustomException(FAILED_GET_LOCK);
+      }
+      // 유저 조회
+      User findUser = getUser(userId);
 
-    // user_chatroom 현재 인원 카운트
-    Long currentUserCount = userChatRoomRepository.countByChatRoomId(roomId);
+      // room 조회
+      ChatRoom chatRoom = chatRoomRepository.findById(roomId) // lock (기존)
+          .orElseThrow(() -> new CustomException(NOT_FOUND_ROOM));
 
-    List<Long> userChatRoomByChatRoomId = userChatRoomRepository.findUserChatRoomByChatRoom_Id(
-        roomId);
+      // user_chatroom 현재 인원 카운트 (비즈니스 로직)
+      Long currentUserCount = userChatRoomRepository.countNonLockByChatRoomId(roomId); // lock (기존)
 
-    if (userChatRoomByChatRoomId.contains(userId)) {
-      throw new CustomException(ErrorCode.ALREADY_JOIN_ROOM);
+      if (chatRoom.getIsPrivate() && chatRoomRequest.getPassword() == null) {
+        throw new CustomException(NEED_TO_PASSWORD);
+      }
+      if (chatRoom.getIsPrivate() && !chatRoomRequest.getPassword().equals(chatRoom.getPassword())) {
+        throw new CustomException(ROOM_PASSWORD_MISMATCH);
+      }
+
+      if (!chatRoom.getIsPrivate() && chatRoom.getPassword() == null) {
+        List<Long> userList = userChatRoomRepository
+            .findUserChatRoomByChatRoom_Id(roomId);
+        if (userList.contains(userId)) {
+          throw new CustomException(ALREADY_JOIN_ROOM);
+        }
+
+        // chatroom 입장
+        if (currentUserCount >= chatRoom.getUserCountMax()) {
+          throw new CustomException(ErrorCode.ROOM_USER_FULL);
+        }
+        // 비밀번호 확인
+
+        UserChatRoom userChatRoom = UserChatRoom.builder()
+            .user(findUser)
+            .chatRoom(chatRoom)
+            .joinDt(now())
+            .build();
+        UserChatRoom save = userChatRoomRepository.save(userChatRoom);
+        String topicName = BASIC_TOPIC + save.getChatRoom().getId();
+        kafkaTemplate.send(topicName, "Subscribed"); // 개선점
+        // 비즈니스 로직 끝
+      }
+    } catch (InterruptedException e) {
+      throw new CustomException(FAILED_GET_LOCK);
+    } finally {
+      lock.unlock();
     }
-
-    // chatroom 입장
-    if (currentUserCount >= chatRoom.getUserCountMax()) {
-      throw new CustomException(ErrorCode.ROOM_USER_FULL);
-    }
-
-    UserChatRoom userChatRoom = UserChatRoom.builder()
-        .user(findUser)
-        .chatRoom(chatRoom)
-        .build();
-    userChatRoomRepository.save(userChatRoom);
-
   }
 
   // 채팅방 전체 조회
-
   @Override
-  @Transactional
+  @Transactional(readOnly = true)
   public List<ChatRoomDto> getRoomList(Pageable pageable) {
     Page<ChatRoom> all = chatRoomRepository.findAll(pageable);
     return getChatRoomDtos(all);
   }
-  // 자신이 생성한 방 리스트 조회
 
+  // 좋아요 순으로 정렬 후 방 전체 조회
+  @Override
+  public List<ChatRoomDto> getRoomHeartSortList(Pageable pageable) {
+    Page<ChatRoom> all = chatRoomRepository.findAllByOrderByHearts(pageable);
+    List<ChatRoom> chatRooms = all.toList();
+    List<ChatRoomDto> chatRoomDtos = new ArrayList<>();
+    for (ChatRoom chatRoom : chatRooms) {
+      ChatRoomDto build = ChatRoomDto.builder()
+          .id(chatRoom.getId())
+          .title(chatRoom.getTitle())
+          .userCountMax(chatRoom.getUserCountMax())
+          .currentUserCount((long) chatRoom.getUserChatRooms().size())
+          .heartCount(chatRoom.getHearts().size())
+          .build();
+      chatRoomDtos.add(build);
+    }
+    return chatRoomDtos;
+  }
+
+  // 자신이 생성한 방 리스트 조회
   @Override
   public List<ChatRoomDto> roomsByCreatorUser(Long userId, Pageable pageable) {
     Page<ChatRoom> userCreateAll = chatRoomRepository.findAllByUserId(userId, pageable);
@@ -117,13 +203,25 @@ public class ChatRoomServiceImpl implements ChatRoomService {
         .findAllByUserChatRoomsUserId(userId, pageable);
     return getChatRoomDtos(userPartAll);
   }
+
   @Override
   @Transactional
   public void outRoom(Long userId, Long roomId) {
     ChatRoom room = getChatRoom(roomId);
+    List<UserChatRoom> userByChatRoomId = userChatRoomRepository
+        .findUserByChatRoomId(roomId);
+    List<Long> userIds = new ArrayList<>();
+    for (UserChatRoom userChatRoom : userByChatRoomId) {
+      Long id = userChatRoom.getUser().getId();
+      userIds.add(id);
+    }
+    // 만약 방에 없는데 나가기를 시도한 경우
+    if (!userIds.contains(userId)) {
+      throw new CustomException(INVALID_REQUEST);
+    }
     // 방장이 아니라면
     if (!Objects.equals(room.getRoomCreator(), userId)) {
-      userChatRoomRepository.deleteUserChatRoomByUserId(userId);
+      userChatRoomRepository.deleteUserChatRoomByUserId(userId); // point select???
       return;
     }
     // 방장이라면 방 삭제
@@ -137,24 +235,121 @@ public class ChatRoomServiceImpl implements ChatRoomService {
   public void deleteRoom(Long userId, Long roomId) {
     ChatRoom room = getChatRoom(roomId);
     if (!Objects.equals(room.getRoomCreator(), userId)) {
-      throw new CustomException(ErrorCode.NOT_ROOM_CREATOR);
+      throw new CustomException(NOT_ROOM_CREATOR);
     }
     chatMsgRepository.deleteChatMsgByChatRoom_Id(roomId);
     userChatRoomRepository.deleteUserChatRoomByChatRoom_Id(roomId);
     chatRoomRepository.deleteById(roomId);
   }
 
-  private User getFindUser(Long userId) {
-    User findUser = userRepository.findById(userId)
-        .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_USER));
-    return findUser;
+  @Override
+  public void updateRoom(Long roomId, ChatRoomUpdateRequest chatRoomUpdateRequest, Long userId) {
+    ChatRoom room = getChatRoom(roomId);
+    String currentRoomTitle = room.getTitle();
+    if (!room.getRoomCreator().equals(userId)) {
+      throw new CustomException(NOT_ROOM_CREATOR);
+    }
+    if (currentRoomTitle.equals(chatRoomUpdateRequest.getTitle())) {
+      throw new CustomException(REQUEST_SAME_AS_CURRENT_TITLE);
+    }
+    room.update(chatRoomUpdateRequest.getTitle(), now());
+    chatRoomRepository.save(room);
+  }
+
+  @Override
+  @Transactional(readOnly = true)
+  public List<ChatRoomUserDto> getRoomUsers(Long roomId, Long userId) {
+    // 방 정보
+    getChatRoom(roomId);
+    // 로그인 유저 정보
+    getUser(userId);
+    // 방에 있는 유저 정보
+    List<UserChatRoom> userIds = userChatRoomRepository
+        .findUserChatRoomByChatRoomId(roomId);
+    // 방에 있지 않은 유저는 볼 수 없음
+    List<Long> userIdList = new ArrayList<>();
+    for (UserChatRoom chatRoom : userIds) {
+      Long id = chatRoom.getUser().getId();
+      userIdList.add(id);
+    }
+    if (!userIdList.contains(userId)) {
+      throw new CustomException(NOT_ROOM_MEMBER);
+    }
+    // DTO 담기
+    List<ChatRoomUserDto> chatRoomUserDtos = new ArrayList<>();
+    for (UserChatRoom userChatRoom : userIds) {
+      ChatRoomUserDto build = ChatRoomUserDto.builder()
+          .nickname(userChatRoom.getUser().getNickname())
+          .status(userChatRoom.getUser().getStatus())
+          .build();
+      chatRoomUserDtos.add(build);
+    }
+    return chatRoomUserDtos;
+  }
+
+  @Override
+  @Transactional(readOnly = true)
+  public List<ChatRoomDto> myHeartRoomList(Long userId, Pageable pageable) {
+    getUser(userId);
+    List<Heart> heart = heartRepository.findByUserId(userId);
+    if (heart.isEmpty()) {
+      throw new CustomException(NOT_FOUND_HEART);
+    }
+    List<Long> ids = new ArrayList<>();
+    for (Heart heart1 : heart) {
+      Long id = heart1.getChatRoom().getId();
+      ids.add(id);
+    }
+    Page<ChatRoom> chatRoomByInId = chatRoomRepository.findChatRoomByInId(ids, pageable);
+    List<ChatRoom> chatRoomList = chatRoomByInId.toList();
+    List<ChatRoomDto> chatRoomDtos = new ArrayList<>();
+    for (ChatRoom chatRoom : chatRoomList) {
+      ChatRoomDto build = ChatRoomDto.builder()
+          .id(chatRoom.getId())
+          .title(chatRoom.getTitle())
+          .heartCount(chatRoom.getHearts().size())
+          .currentUserCount((long) chatRoom.getUserChatRooms().size())
+          .userCountMax(chatRoom.getUserCountMax())
+          .build();
+      chatRoomDtos.add(build);
+    }
+    return chatRoomDtos;
+  }
+
+  @Override
+  @Transactional(readOnly = true)
+  public List<ChatRoomDto> searchRoomByTitle(String roomName, Long userId, Pageable pageable) {
+    getUser(userId);
+    Page<ChatRoom> search = chatRoomRepository.findByTitleContaining(roomName, pageable);
+
+    List<ChatRoom> searchRoomList = search.toList();
+    List<ChatRoomDto> chatRoomDtos = new ArrayList<>();
+    if (searchRoomList.size() == 0) {
+      throw new CustomException(NOT_FOUND_ROOM);
+    }
+
+    for (ChatRoom chatRoom : searchRoomList) {
+      ChatRoomDto build = ChatRoomDto.builder()
+          .id(chatRoom.getId())
+          .title(chatRoom.getTitle())
+          .heartCount(chatRoom.getHearts().size())
+          .currentUserCount((long) chatRoom.getUserChatRooms().size())
+          .userCountMax(chatRoom.getUserCountMax())
+          .build();
+      chatRoomDtos.add(build);
+    }
+    return chatRoomDtos;
+  }
+
+  private User getUser(Long userId) {
+    return userRepository.findById(userId)
+        .orElseThrow(() -> new CustomException(NOT_FOUND_USER));
   }
 
   private ChatRoom getChatRoom(Long roomId) {
-    ChatRoom room = chatRoomRepository
+    return chatRoomRepository
         .findChatRoomById(roomId)
-        .orElseThrow(() -> new CustomException(ErrorCode.NONE_ROOM));
-    return room;
+        .orElseThrow(() -> new CustomException(NONE_ROOM));
   }
 
   // 방 조회 DTO 변환 메서드 추출
